INTYGFV-16762: Updated to int instead of long.

+++ certificate-service/app/src/main/java/se/inera/intyg/certificateservice/application/unit/dto/UnitStatisticsDTO.java
@@ -10,9 +10,9 @@ import se.inera.intyg.certificateservice.application.unit.dto.UnitStatisticsDTO.
 @Value
 @Builder
 public class UnitStatisticsDTO {
-  
-  long draftCount;
-  long unhandledMessageCount;
+
+  int draftCount;
+  int unhandledMessageCount;
 
   @JsonPOJOBuilder(withPrefix = "")
   public static class UnitStatisticsDTOBuilder {
